requirements.txt
flask
flask-restful 
flask-sqlalchemy 
python-dotenv
requests
flasgger